Renamed package org.mule.tooling.restsdk.wizzard to org.mule.tooling.restsdk.wizard.
