+++ data-weave-plugin/src/main/java/org/mule/tooling/restsdk/wizard/RestSdkConfigurationModel.java
@@ -1,4 +1,4 @@
-package org.mule.tooling.restsdk.wizzard;
+package org.mule.tooling.restsdk.wizard;
 
 public class RestSdkConfigurationModel {
     private String restSdkVersion;
