+++ data-weave-plugin/src/main/resources/META-INF/plugin.xml
@@ -619,7 +619,7 @@
 
         <psi.referenceContributor language="yaml"
                                   implementation="org.mule.tooling.restsdk.reference.RestSdkReferenceContributor"/>
-        <moduleBuilder builderClass="org.mule.tooling.restsdk.wizzard.RestSdkModuleBuilder"/>
+        <moduleBuilder builderClass="org.mule.tooling.restsdk.wizard.RestSdkModuleBuilder"/>
         <completion.contributor language="yaml"
                                 implementationClass="org.mule.tooling.restsdk.completion.RestSdkCompletionContributor"/>
         <annotator language="yaml" implementationClass="org.mule.tooling.restsdk.annotator.RestSdkAnnotator"/>
